@@ -189,48 +189,27 @@ Value Schema::_load_object(const std::string& name, const std::string& type) con
   Value ret_val;
   try
   {
-    boost::shared_ptr<BaseSession> sess(boost::dynamic_pointer_cast<BaseSession>(_session.lock()));
-    if (sess)
+    std::string found_type(type);
+    if (_session.lock()->db_object_exists(found_type, name, _name))
     {
-      shcore::Argument_list args;
-      args.push_back(Value(_name));
-      args.push_back(Value(name));
-
-      Value myres = sess->executeAdminCommand("list_objects", args);
-      boost::shared_ptr<mysh::mysqlx::Resultset> my_res = myres.as_object<mysh::mysqlx::Resultset>();
-
-      Value raw_entry = my_res->next(shcore::Argument_list());
-
-      if (raw_entry)
+      if (found_type == "TABLE")
       {
-        boost::shared_ptr<mysh::Row> row = raw_entry.as_object<mysh::Row>();
-        std::string object_name = row->get_member("name").as_string();
-        std::string object_type = row->get_member("type").as_string();
-
-        if (type.empty() || type == object_type)
-        {
-          if (object_type == "TABLE")
-          {
-            boost::shared_ptr<Table> table(new Table(shared_from_this(), object_name));
-            ret_val = Value(boost::static_pointer_cast<Object_bridge>(table));
-            (*_tables)[name] = ret_val;
-          }
-          else if (object_type == "VIEW")
-          {
-            boost::shared_ptr<View> view(new View(shared_from_this(), object_name));
-            ret_val = Value(boost::static_pointer_cast<Object_bridge>(view));
-            (*_views)[name] = ret_val;
-          }
-          else if (object_type == "COLLECTION")
-          {
-            boost::shared_ptr<Collection> collection(new Collection(shared_from_this(), object_name));
-            ret_val = Value(boost::static_pointer_cast<Object_bridge>(collection));
-            (*_collections)[name] = ret_val;
-          }
-        }
+        boost::shared_ptr<Table> table(new Table(shared_from_this(), name));
+        ret_val = Value(boost::static_pointer_cast<Object_bridge>(table));
+        (*_tables)[name] = ret_val;
       }
-
-      my_res->all(shcore::Argument_list());
+      else if (found_type == "VIEW")
+      {
+        boost::shared_ptr<View> view(new View(shared_from_this(), name));
+        ret_val = Value(boost::static_pointer_cast<Object_bridge>(view));
+        (*_views)[name] = ret_val;
+      }
+      else if (found_type == "COLLECTION")
+      {
+        boost::shared_ptr<Collection> collection(new Collection(shared_from_this(), name));
+        ret_val = Value(boost::static_pointer_cast<Object_bridge>(collection));
+        (*_collections)[name] = ret_val;
+      }
     }
   }
   CATCH_AND_TRANSLATE();
@@ -246,12 +225,11 @@ Value Schema::_load_object(const std::string& name, const std::string& type) con
 * \param name the name of the table to retrieve.
 * \return the Table object or undefined.
 */
-Table Schema::getTable(String name)
-{}
+Table Schema::getTable(String name){}
 #endif
 shcore::Value Schema::getTable(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, (class_name() + "::getTable").c_str());
+  args.ensure_count(1, (class_name() + ".getTable").c_str());
 
   std::string name = args.string_at(0);
 
@@ -266,12 +244,11 @@ shcore::Value Schema::getTable(const shcore::Argument_list &args)
 * \param name the name of the collection to retrieve.
 * \return the Collection object or undefined.
 */
-CollectionRef Schema::getCollection(String name)
-{}
+Collection Schema::getCollection(String name){}
 #endif
 shcore::Value Schema::getCollection(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, (class_name() + "::getCollection").c_str());
+  args.ensure_count(1, (class_name() + ".getCollection").c_str());
 
   std::string name = args.string_at(0);
 
@@ -286,12 +263,11 @@ shcore::Value Schema::getCollection(const shcore::Argument_list &args)
 * \param name the name of the view to retrieve.
 * \return the View object or undefined.
 */
-View Schema::getView(String name)
-{}
+View Schema::getView(String name){}
 #endif
 shcore::Value Schema::getView(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, (class_name() + "::getCollection").c_str());
+  args.ensure_count(1, (class_name() + ".getCollection").c_str());
 
   std::string name = args.string_at(0);
 
@@ -303,17 +279,16 @@ shcore::Value Schema::getView(const shcore::Argument_list &args)
 * Creates in the current schema a new collection with the specified name and retrieves an object representing the new collection created.
 * TODO: What are the valid identifiers for collection names?
 * \sa getCollections(), getCollection()
-* \param name the name of the colllection.
+* \param name the name of the collection.
 * \return the new created collection.
 */
-CollectionRef Schema::createCollection(String name)
-{}
+Collection Schema::createCollection(String name){}
 #endif
 shcore::Value Schema::createCollection(const shcore::Argument_list &args)
 {
   Value ret_val;
 
-  args.ensure_count(1, (class_name() + "::createCollection").c_str());
+  args.ensure_count(1, (class_name() + ".createCollection").c_str());
 
   // Creates the collection on the server
   shcore::Argument_list command_args;
@@ -323,7 +298,7 @@ shcore::Value Schema::createCollection(const shcore::Argument_list &args)
   boost::shared_ptr<BaseSession> sess(boost::static_pointer_cast<BaseSession>(_session.lock()));
   sess->executeAdminCommand("create_collection", command_args);
 
-  // If this is reached it imlies all went OK on the previous operation
+  // If this is reached it implies all went OK on the previous operation
   std::string name = args.string_at(0);
   boost::shared_ptr<Collection> collection(new Collection(shared_from_this(), name));
   ret_val = Value(boost::static_pointer_cast<Object_bridge>(collection));
