@@ -34,6 +34,7 @@
 #include <boost/lexical_cast.hpp>
 #include <boost/format.hpp>
 #include <boost/pointer_cast.hpp>
+#include <boost/algorithm/string.hpp>
 
 #define MAX_COLUMN_LENGTH 1024
 #define MIN_COLUMN_LENGTH 4
@@ -42,7 +43,7 @@ using namespace mysh;
 using namespace shcore;
 using namespace mysh::mysqlx;
 
-REGISTER_OBJECT(mysqlx, Session);
+REGISTER_OBJECT(mysqlx, XSession);
 REGISTER_OBJECT(mysqlx, NodeSession);
 REGISTER_OBJECT(mysqlx, Expression);
 
@@ -50,12 +51,12 @@ REGISTER_OBJECT(mysqlx, Expression);
 
 #ifdef DOXYGEN
 /**
-* Creates a Session instance using the provided connection data.
+* Creates a XSession instance using the provided connection data.
 * \param connectionData the connection string used to connect to the database.
 * \param password if provided, will override the password in the connection string.
-* \return An Session instance.
+* \return An XSession instance.
 *
-* A Session object uses the X Protocol to allow executing operations on the connected MySQL Product.
+* A XSession object uses the X Protocol to allow executing operations on the connected MySQL Product.
 *
 * The format of the connection string can be as follows for connections using the TCP protocol:
 *
@@ -63,15 +64,11 @@ REGISTER_OBJECT(mysqlx, Expression);
 *
 * or as follows for connections using a socket or named pipe:
 *
-* [user[:pass\@] ::socket[/db]
+* [user[:pass\@]\::socket[/db]
 *
-* \sa Session
+* \sa XSession
 */
-Session getSession(String connectionData, String password)
-{
-  Session a;
-  return a;
-}
+XSession getSession(String connectionData, String password){}
 
 /**
 * This function works the same as the function above, except that the connection data comes enclosed on a dictionary object.
@@ -82,33 +79,21 @@ Session getSession(String connectionData, String password)
 *  - dbUser, the user to authenticate against.
 *  - dbPassword, the password of the user user to authenticate against.
 * \param password if provided, will override the password in the connection string.
-* \return An Session instance.
+* \return An XSession instance.
 *
-* \sa Session
+* \sa XSession
 */
-Session getSession(Map connectionData, String password)
-{
-  Session a;
-  return a;
-}
+XSession getSession(Map connectionData, String password){}
 
 /**
 * This function works the same as getSession(String connectionData, String password), except that it will return a NodeSession object which allows SQL Execution.
 */
-NodeSession getNodeSession(String connectionData, String password)
-{
-  NodeSession a;
-  return a;
-}
+NodeSession getNodeSession(String connectionData, String password){}
 
 /**
 * This function works the same as getSession(Map connectionData, String password), except that it will return a NodeSession object which allows SQL Execution.
 */
-NodeSession getNodeSession(Map connectionData, String password)
-{
-  NodeSession a;
-  return a;
-}
+NodeSession getNodeSession(Map connectionData, String password){}
 
 #endif
 
@@ -191,11 +176,9 @@ Value BaseSession::connect(const Argument_list &args)
 * \brief Closes the session.
 * After closing the session it is still possible to make read only operation to gather metadata info, like getTable(name) or getSchemas().
 */
-Undefined BaseSession::close()
-{}
+Undefined BaseSession::close(){}
 #endif
-
-Value BaseSession::close(const Argument_list &args)
+Value BaseSession::close(const shcore::Argument_list &args)
 {
   std::string function_name = class_name() + ".close";
 
@@ -206,12 +189,6 @@ Value BaseSession::close(const Argument_list &args)
   return shcore::Value();
 }
 
-/**
-* Creates a SqlExecute object to allow running the received SQL statement on the target MySQL Server.
-* \param statement A string containing the SQL statement to be executed.
-* \return A Resultset object to allow pulling the statement result data.
-* \sa SqlExecute, Resultset
-*/
 Value BaseSession::sql(const Argument_list &args)
 {
   std::string function_name = class_name() + ".sql";
@@ -224,6 +201,38 @@ Value BaseSession::sql(const Argument_list &args)
   try
   {
     ret_val = executeSql(args.string_at(0), shcore::Argument_list());
+  }
+  CATCH_AND_TRANSLATE();
+
+  return ret_val;
+}
+
+#ifdef DOXYGEN
+/**
+* Creates a schema on the database and returns the corresponding object.
+* \param name A string value indicating the schema name.
+* \return The created schema object.
+* \exception An exception is thrown if an error occurs creating the XSession.
+*/
+Schema BaseSession::createSchema(String name){}
+#endif
+Value BaseSession::createSchema(const shcore::Argument_list &args)
+{
+  std::string function_name = class_name() + ".createSchema";
+  args.ensure_count(1, function_name.c_str());
+
+  Value ret_val;
+  try
+  {
+    std::string schema = args.string_at(0);
+    std::string statement = "create schema " + schema;
+    ret_val = executeStmt("sql", statement, shcore::Argument_list());
+
+    // if reached this point it indicates that there were no errors
+    boost::shared_ptr<Schema> object(new Schema(_get_shared_this(), schema));
+    ret_val = shcore::Value(boost::static_pointer_cast<Object_bridge>(object));
+
+    (*_schemas)[schema] = ret_val;
   }
   CATCH_AND_TRANSLATE();
 
@@ -336,24 +345,23 @@ bool BaseSession::has_member(const std::string &prop) const
 
 #ifdef DOXYGEN
 /**
-* \brief Retrieves the Schema configured as default for the session, if none, returns Null.
+* Retrieves the Schema configured as default for the session.
+* \return A Schema object or Null
 */
-Schema BaseSession::getDefaultSchema()
-{}
+Schema BaseSession::getDefaultSchema(){}
 
 /**
-* \brief Retrieves the List of Schemas available on the session.
+* Retrieves the Schemas available on the session.
+* \return A Map containing the Schema objects available o the session.
 */
-List BaseSession::getSchemas()
-{}
+Map BaseSession::getSchemas(){}
 
 /**
-* \brief Retrieves the connectionData in string format.
+* Retrieves the connection data for this session in string format.
+* \return A string representing the connection data.
 */
-String BaseSession::getUri()
-{}
+String BaseSession::getUri(){}
 #endif
-
 Value BaseSession::get_member(const std::string &prop) const
 {
   // Retrieves the member first from the parent
@@ -378,12 +386,15 @@ Value BaseSession::get_member(const std::string &prop) const
   }
   else
   {
-    // Since the property was not satisfied, we assume it is a schema and
-    // proceed to retrieve it
-    shcore::Argument_list args;
-    args.push_back(Value(prop));
+    if (_schemas->has_key(prop))
+    {
+      boost::shared_ptr<Schema> schema = (*_schemas)[prop].as_object<Schema>();
 
-    ret_val = get_schema(args);
+      // This will validate the schema continues valid
+      schema->cache_table_objects();
+
+      ret_val = (*_schemas)[prop];
+    }
   }
 
   return ret_val;
@@ -455,6 +466,17 @@ void BaseSession::_load_schemas()
   CATCH_AND_TRANSLATE();
 }
 
+#ifdef DOXYGEN
+/**
+* Retrieves a Schema object from the current session through it's name.
+* \param name The name of the Schema object to be retrieved.
+* \return The Schema object with the given name.
+* \exception An exception is thrown if the given name is not a valid schema on the XSession.
+* \sa Schema
+*/
+Schema BaseSession::getSchema(String name){}
+#endif
+
 shcore::Value BaseSession::get_schema(const shcore::Argument_list &args) const
 {
   std::string function_name = class_name() + ".getSchema";
@@ -481,14 +503,12 @@ shcore::Value BaseSession::get_schema(const shcore::Argument_list &args) const
       (*_schemas)[name] = Value(boost::static_pointer_cast<Object_bridge>(schema));
     }
     else
-      throw Exception::runtime_error("Session not connected");
+      throw Exception::runtime_error("XSession not connected");
   }
 
   // If this point is reached, the schema will be there!
   return (*_schemas)[name];
 }
-
-Schema setDefaultSchema();
 
 #ifdef DOXYGEN
 /**
@@ -500,8 +520,7 @@ Schema setDefaultSchema();
 * \param name the name of the new schema to switch to.
 * \return the Schema object for the new schema.
 */
-Schema BaseSession::setDefaultSchema(String name)
-{}
+Schema BaseSession::setDefaultSchema(String name){}
 #endif
 shcore::Value BaseSession::set_default_schema(const shcore::Argument_list &args)
 {
@@ -521,18 +540,16 @@ shcore::Value BaseSession::set_default_schema(const shcore::Argument_list &args)
     _update_default_schema(name);
   }
   else
-    throw Exception::runtime_error("Session not connected");
+    throw Exception::runtime_error("XSession not connected");
 
   return get_member("defaultSchema");
 }
-
-//Undefined setFetchWarnings(Bool value);
 
 shcore::Value BaseSession::set_fetch_warnings(const shcore::Argument_list &args)
 {
   Value ret_val;
 
-  args.ensure_count(1, (class_name() + "::setFetchWarnings").c_str());
+  args.ensure_count(1, (class_name() + ".setFetchWarnings").c_str());
 
   bool enable = args.bool_at(0);
   std::string command = enable ? "enable_notices" : "disable_notices";
@@ -560,16 +577,86 @@ void BaseSession::_update_default_schema(const std::string& name)
   }
 }
 
-boost::shared_ptr<BaseSession> Session::_get_shared_this() const
+void BaseSession::drop_db_object(const std::string &type, const std::string &name, const std::string& owner)
 {
-  boost::shared_ptr<const Session> shared = shared_from_this();
+  if (type == "Schema")
+    executeStmt("sql", "drop schema `" + name + "`", shcore::Argument_list());
+  else if (type == "View")
+    executeStmt("sql", "drop view `" + owner + "`.`" + name + "`", shcore::Argument_list());
+  else
+  {
+    shcore::Argument_list command_args;
+    command_args.push_back(Value(owner));
+    command_args.push_back(Value(name));
 
-  return boost::const_pointer_cast<Session>(shared);
+    executeAdminCommand("drop_collection", command_args);
+  }
 }
 
-boost::shared_ptr<shcore::Object_bridge> Session::create(const shcore::Argument_list &args)
+/*
+* This function verifies if the given object exist in the database, works for schemas, tables, views and collections.
+* The check for tables, views and collections is done is done based on the type.
+* If type is not specified and an object with the name is found, the type will be returned.
+*/
+bool BaseSession::db_object_exists(std::string &type, const std::string &name, const std::string& owner)
 {
-  boost::shared_ptr<Session> session(new Session());
+  std::string statement;
+  bool ret_val = false;
+
+  if (type == "Schema")
+  {
+    shcore::Value res = executeStmt("sql", "show databases like \"" + name + "\"", shcore::Argument_list());
+    boost::shared_ptr<Resultset> res_obj = res.as_object<Resultset>();
+
+    ret_val = res_obj->all(shcore::Argument_list()).as_array()->size() > 0;
+  }
+  else
+  {
+    shcore::Argument_list args;
+    args.push_back(Value(owner));
+    args.push_back(Value(name));
+
+    Value myres = executeAdminCommand("list_objects", args);
+    boost::shared_ptr<mysh::mysqlx::Resultset> my_res = myres.as_object<mysh::mysqlx::Resultset>();
+
+    Value raw_entry = my_res->next(shcore::Argument_list());
+
+    if (raw_entry)
+    {
+      boost::shared_ptr<mysh::Row> row = raw_entry.as_object<mysh::Row>();
+      std::string object_name = row->get_member("name").as_string();
+      std::string object_type = row->get_member("type").as_string();
+
+      if (type.empty())
+      {
+        type = object_type;
+        ret_val = true;
+      }
+      else
+      {
+        boost::algorithm::to_upper(type);
+
+        if (type == object_type)
+          ret_val = true;
+      }
+    }
+
+    my_res->all(shcore::Argument_list());
+  }
+
+  return ret_val;
+}
+
+boost::shared_ptr<BaseSession> XSession::_get_shared_this() const
+{
+  boost::shared_ptr<const XSession> shared = shared_from_this();
+
+  return boost::const_pointer_cast<XSession>(shared);
+}
+
+boost::shared_ptr<shcore::Object_bridge> XSession::create(const shcore::Argument_list &args)
+{
+  boost::shared_ptr<XSession> session(new XSession());
 
   session->connect(args);
 
@@ -599,18 +686,25 @@ boost::shared_ptr<shcore::Object_bridge> NodeSession::create(const shcore::Argum
 
 #ifdef DOXYGEN
 /**
-* Executes an sql statement and returns a Resultset object
-* \param sql The statement to be executed
+* Creates a SqlExecute object to allow running the received SQL statement on the target MySQL Server.
+* \param sql A string containing the SQL statement to be executed.
+* \return A SqlExecute object.
+* \sa SqlExecute
 *
-* \return Resultset object
+* This method creates an SqlExecute object which is a SQL execution handler.
+*
+* The SqlExecute class has functions that allow defining the way the statement will be executed and allows doing parameter binding.
+*
+* The received SQL is set on the execution handler.
 *
 * JavaScript Example
 * \code{.js}
-* var result = session.sql("show databases").execute();
+* var sql = session.sql("select * from mydb.students where  age > ?");
+* var result = sql.bind(18).execute();
 * \endcode
+* \sa SqlExecute
 */
-Resultset NodeSession::sql(String sql)
-{}
+Resultset NodeSession::sql(String sql){}
 #endif
 shcore::Value NodeSession::sql(const shcore::Argument_list &args)
 {
