@@ -40,6 +40,8 @@ DatabaseObject::DatabaseObject(boost::shared_ptr<ShellBaseSession> session, boos
   add_method("getName", boost::bind(&DatabaseObject::get_member_method, this, _1, "getName", "name"), NULL);
   add_method("getSession", boost::bind(&DatabaseObject::get_member_method, this, _1, "getSession", "session"), NULL);
   add_method("getSchema", boost::bind(&DatabaseObject::get_member_method, this, _1, "getSchema", "schema"), NULL);
+  add_method("drop", boost::bind(&DatabaseObject::drop, this, _1), "data");
+  add_method("existInDatabase", boost::bind(&DatabaseObject::existInDatabase, this, _1), "data");
 }
 
 DatabaseObject::~DatabaseObject()
@@ -79,7 +81,7 @@ std::vector<std::string> DatabaseObject::get_members() const
 
 shcore::Value DatabaseObject::get_member_method(const shcore::Argument_list &args, const std::string& method, const std::string& prop)
 {
-  std::string function = class_name() + "::" + method;
+  std::string function = class_name() + "." + method;
   args.ensure_count(0, function.c_str());
 
   return get_member(prop);
@@ -105,6 +107,34 @@ bool DatabaseObject::has_member(const std::string &prop) const
     prop == "schema";
 }
 
+#ifdef DOXYGEN
+/**
+* Returns the name of this database object.
+* \return the name as an String object.
+*/
+String DatabaseObject::getName(){}
+
+/**
+* Returns the Session object of this database object.
+* \return the Session object used to get to this object.
+*
+* Note that the returned object can be any of:
+* - XSession: if the object was created/retrieved using an XSession instance.
+* - NodeSession: if the object was created/retrieved using an NodeSession instance.
+* - ClassicSession: if the object was created/retrieved using an ClassicSession instance.
+*/
+Object DatabaseObject::getSession(){}
+
+/**
+* Returns the Schema object of this database object.
+* \return the object for this schema of this database object.
+*
+* Note that the returned object can be any of:
+* - Schema: if the object was created/retrieved using a Schema instance.
+* - ClassicSchema: if the object was created/retrieved using an ClassicSchema instance.
+*/
+Object DatabaseObject::getSchema(){}
+#endif
 Value DatabaseObject::get_member(const std::string &prop) const
 {
   Value ret_val;
@@ -129,4 +159,29 @@ Value DatabaseObject::get_member(const std::string &prop) const
     ret_val = Cpp_object_bridge::get_member(prop);
 
   return ret_val;
+}
+
+#ifdef DOXYGEN
+/**
+* Drops this object from the database.
+*/
+Undefined DatabaseObject::drop(){}
+#endif
+shcore::Value DatabaseObject::drop(const shcore::Argument_list &args)
+{
+  _session.lock()->drop_db_object(class_name(), _name, _schema._empty() ? "" : _schema.lock()->get_member("name").as_string());
+
+  return shcore::Value();
+}
+
+#ifdef DOXYGEN
+/**
+* Verifies if this object exists in the database.
+*/
+Undefined DatabaseObject::existInDatabase(){}
+#endif
+shcore::Value DatabaseObject::existInDatabase(const shcore::Argument_list &args)
+{
+  std::string cname(class_name());
+  return shcore::Value(_session.lock()->db_object_exists(cname, _name, _schema._empty() ? "" : _schema.lock()->get_member("name").as_string()));
 }
