@@ -165,7 +165,7 @@ boost::shared_ptr<mysh::ShellBaseSession> mysh::connect_session(const shcore::Ar
   switch (session_type)
   {
     case Application:
-      ret_val.reset(new mysh::mysqlx::Session());
+      ret_val.reset(new mysh::mysqlx::XSession());
       break;
     case Node:
       ret_val.reset(new mysh::mysqlx::NodeSession());
@@ -187,6 +187,7 @@ boost::shared_ptr<mysh::ShellBaseSession> mysh::connect_session(const shcore::Ar
 
 ShellBaseSession::ShellBaseSession()
 {
+  add_method("createSchema", boost::bind(&ShellBaseSession::createSchema, this, _1), "name", shcore::String, NULL);
   add_method("getDefaultSchema", boost::bind(&ShellBaseSession::get_member_method, this, _1, "getDefaultSchema", "defaultSchema"), NULL);
   add_method("getSchema", boost::bind(&ShellBaseSession::get_schema, this, _1), "name", shcore::String, NULL);
   add_method("getSchemas", boost::bind(&ShellBaseSession::get_member_method, this, _1, "getSchemas", "schemas"), NULL);
@@ -232,7 +233,7 @@ std::vector<std::string> ShellBaseSession::get_members() const
 
 shcore::Value ShellBaseSession::get_member_method(const shcore::Argument_list &args, const std::string& method, const std::string& prop)
 {
-  std::string function = class_name() + "::" + method;
+  std::string function = class_name() + "." + method;
   args.ensure_count(0, function.c_str());
 
   return get_member(prop);
