@@ -58,39 +58,37 @@ Collection::~Collection()
 #ifdef DOXYGEN
 /**
 * Adds a document to a collection.
-* \param A documents to be added into the collection.
+* \param document The document to be added into the collection.
 * \return A CollectionAdd object.
 *
 * To be added, the document must have a property named '_id' with a universal unique identifier (UUID). If the property is missing, it is set with an auto generated UUID.
 *
-* This method creates a CollectionAdd object which is a document addition handler, the received document is added into this handler.
+* This function creates a CollectionAdd object which is a document addition handler, the received document is added into this handler.
 *
-* The CollectionAdd has other functions that allow specifying the way the addition occurs.
+* The CollectionAdd class has other functions that allow specifying the way the addition occurs.
 *
-* The addition is done when the execute method is called on the handler.
+* The addition is done when the execute function is called on the handler.
 *
 * \sa CollectionAdd
 */
-CollectionAdd Collection::add(Document document)
-{}
+CollectionAdd Collection::add(Document document){}
 
 /**
 * Adds a list of documents to a collection.
-* \param The document list to be added into the collection.
+* \param documents The document list to be added into the collection.
 * \return A CollectionAdd object.
 *
 * Every document to be added must have a property named '_id' with a universal unique identifier (UUID). If the property is missing, it is set with an auto generated UUID.
 *
-* This method creates a CollectionAdd object which is a document addition handler, the received documents are added into this handler.
+* This function creates a CollectionAdd object which is a document addition handler, the received documents are added into this handler.
 *
-* The CollectionAdd has other functions that allow specifying the way the addition occurs.
+* The CollectionAdd class has other functions that allow specifying the way the addition occurs.
 *
-* The addition is done when the execute method is called on the handler.
+* The addition is done when the execute function is called on the handler.
 *
 * \sa CollectionAdd
 */
-CollectionAdd Collection::add(List documents)
-{}
+CollectionAdd Collection::add(List documents){}
 #endif
 shcore::Value Collection::add_(const shcore::Argument_list &args)
 {
@@ -101,20 +99,19 @@ shcore::Value Collection::add_(const shcore::Argument_list &args)
 
 #ifdef DOXYGEN
 /**
-* Updates documents in a collection.
+* Creates a collection update handler.
 * \param searchCondition An optional string with the filter expression of the documents to be modified.
 * \return A CollectionFind object.
 *
-* This method creates a CollectionModify object which is a document update handler.
+* This function creates a CollectionModify object which is a document update handler.
 *
-* The CollectionModify object has several functions that allow specifying the way the update occurs, if a searchCondition was specified, it will be set on the handler.
+* The CollectionModify class has several functions that allow specifying the way the update occurs, if a searchCondition was specified, it will be set on the handler.
 *
-* The update is done when the execute method is called on the handler.
+* The update is done when the execute function is called on the handler.
 *
 * \sa CollectionModify
 */
-CollectionModify Collection::modify(String searchCondition)
-{}
+CollectionModify Collection::modify(String searchCondition){}
 #endif
 shcore::Value Collection::modify_(const shcore::Argument_list &args)
 {
@@ -125,20 +122,19 @@ shcore::Value Collection::modify_(const shcore::Argument_list &args)
 
 #ifdef DOXYGEN
 /**
-* Removes documents from a collection.
-* \param searchCondition An optional string with the filter expression of the documents to be removed.
+* Creates a document deletion handler.
+* \param searchCondition An optional string with the filter expression of the documents to be deleted.
 * \return A CollectionRemove object.
 *
-* This method creates a CollectionRemove object which is a document removal handler.
+* This function creates a CollectionRemove object which is a document deletion handler.
 *
-* The CollectionRemove object has several functions that allow specifying what should be deleted and how, if a searchCondition was specified, it will be set on the handler.
+* The CollectionRemove class has several functions that allow specifying what should be deleted and how, if a searchCondition was specified, it will be set on the handler.
 *
-* The removal is done when the execute method is called on the handler.
+* The deletion is done when the execute function is called on the handler.
 *
 * \sa CollectionRemove
 */
-CollectionRemove Collection::remove(String searchCondition)
-{}
+CollectionRemove Collection::remove(String searchCondition){}
 #endif
 shcore::Value Collection::remove_(const shcore::Argument_list &args)
 {
@@ -150,19 +146,18 @@ shcore::Value Collection::remove_(const shcore::Argument_list &args)
 #ifdef DOXYGEN
 /**
 * Retrieves documents from a collection.
-* \param searchCondition An optional string with the filter expression of the documents to be retrieved.
+* \param searchCriteria An optional string with the filter expression of the documents to be retrieved.
 * \return A CollectionFind object.
 *
-* This method creates a CollectionFind object which is a document selection handler.
+* This function creates a CollectionFind object which is a document selection handler.
 *
-* The CollectionFind object has several functions that allow specifying what should be retrieved from the collection, if a searchCondition was specified, it will be set on the handler.
+* The CollectionFind class has several functions that allow specifying what should be retrieved from the collection, if a searchCondition was specified, it will be set on the handler.
 *
-* The selection will be returned when the execute method is called on the handler.
+* The selection will be returned when the execute function is called on the handler.
 *
 * \sa CollectionFind
 */
-CollectionFind Collection::find(String searchCriteria)
-{}
+CollectionFind Collection::find(String searchCriteria){}
 #endif
 shcore::Value Collection::find_(const shcore::Argument_list &args)
 {
