@@ -40,14 +40,13 @@ namespace mysh
     class Schema;
 
     /**
-    * Represents a Collection on an Schema, retrieved with a session created using the X Protocol.
+    * Represents a Collection on an Schema, retrieved with session created using the mysqlx module.
     * \todo Implement and document as()
     * \todo Implement and document createIndex()
     * \todo Implement and document dropIndex()
     * \todo Implement and document getIndexes()
     * \todo Implement and document newDoc()
     * \todo Implement and document count()
-    * \todo Implement and document drop()
     */
     class Collection : public DatabaseObject, public boost::enable_shared_from_this<Collection>
     {
