@@ -80,13 +80,12 @@ CollectionModify::CollectionModify(boost::shared_ptr<Collection> owner)
 * \sa Usage examples at execute(ExecuteOptions options).
 * \sa Collection
 */
-CollectionModify CollectionModify::modify(String searchCondition)
-{}
+CollectionModify CollectionModify::modify(String searchCondition){}
 #endif
 shcore::Value CollectionModify::modify(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(0, 1, "CollectionModify::modify");
+  args.ensure_count(0, 1, "CollectionModify.modify");
 
   boost::shared_ptr<Collection> collection(boost::static_pointer_cast<Collection>(_owner.lock()));
 
@@ -103,7 +102,7 @@ shcore::Value CollectionModify::modify(const shcore::Argument_list &args)
       // Updates the exposed functions
       update_functions("modify");
     }
-    CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::modify");
+    CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.modify");
   }
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
@@ -122,7 +121,7 @@ shcore::Value CollectionModify::modify(const shcore::Argument_list &args)
 *
 * The attribute addition will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -148,13 +147,12 @@ shcore::Value CollectionModify::modify(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::set(String attribute, Value value)
-{}
+CollectionModify CollectionModify::set(String attribute, Value value){}
 #endif
 shcore::Value CollectionModify::set(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(2, "CollectionModify::set");
+  args.ensure_count(2, "CollectionModify.set");
 
   try
   {
@@ -163,7 +161,7 @@ shcore::Value CollectionModify::set(const shcore::Argument_list &args)
 
     update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::set");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.set");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -179,7 +177,7 @@ shcore::Value CollectionModify::set(const shcore::Argument_list &args)
 *
 * The attribute removal will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -205,8 +203,7 @@ shcore::Value CollectionModify::set(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::unset(String attribute)
-{}
+CollectionModify CollectionModify::unset(String attribute){}
 
 /**
 * Removes attributes from documents in a collection.
@@ -218,7 +215,7 @@ CollectionModify CollectionModify::unset(String attribute)
 *
 * The attribute removal will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -244,13 +241,12 @@ CollectionModify CollectionModify::unset(String attribute)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::unset(List attributes)
-{}
+CollectionModify CollectionModify::unset(List attributes){}
 #endif
 shcore::Value CollectionModify::unset(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_at_least(1, "CollectionModify::unset");
+  args.ensure_at_least(1, "CollectionModify.unset");
 
   try
   {
@@ -300,7 +296,7 @@ shcore::Value CollectionModify::unset(const shcore::Argument_list &args)
     if (unset_count)
       update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::unset");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.unset");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -316,7 +312,7 @@ shcore::Value CollectionModify::unset(const shcore::Argument_list &args)
 *
 * The attribute addition will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -342,13 +338,12 @@ shcore::Value CollectionModify::unset(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::merge(Document document)
-{}
+CollectionModify CollectionModify::merge(Document document){}
 #endif
 shcore::Value CollectionModify::merge(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(1, "CollectionModify::merge");
+  args.ensure_count(1, "CollectionModify.merge");
 
   try
   {
@@ -360,7 +355,7 @@ shcore::Value CollectionModify::merge(const shcore::Argument_list &args)
 
     update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::merge");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.merge");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -376,7 +371,7 @@ shcore::Value CollectionModify::merge(const shcore::Argument_list &args)
 *
 * The insertion of the value will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -402,13 +397,12 @@ shcore::Value CollectionModify::merge(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::arrayInsert(String path, Value value)
-{}
+CollectionModify CollectionModify::arrayInsert(String path, Value value){}
 #endif
 shcore::Value CollectionModify::array_insert(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(2, "CollectionModify::arrayInsert");
+  args.ensure_count(2, "CollectionModify.arrayInsert");
 
   try
   {
@@ -417,7 +411,7 @@ shcore::Value CollectionModify::array_insert(const shcore::Argument_list &args)
     // Updates the exposed functions
     update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::arrayInsert");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.arrayInsert");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -431,7 +425,7 @@ shcore::Value CollectionModify::array_insert(const shcore::Argument_list &args)
 *
 * Adds an opertion into the modify handler to append a value into an array attribute on the documents that were included on the selection filter and limit.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -459,13 +453,12 @@ shcore::Value CollectionModify::array_insert(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::arrayAppend(String path, Value value)
-{}
+CollectionModify CollectionModify::arrayAppend(String path, Value value){}
 #endif
 shcore::Value CollectionModify::array_append(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(2, "CollectionModify::arrayAppend");
+  args.ensure_count(2, "CollectionModify.arrayAppend");
 
   try
   {
@@ -475,7 +468,7 @@ shcore::Value CollectionModify::array_append(const shcore::Argument_list &args)
 
     update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::arrayAppend");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.arrayAppend");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -490,7 +483,7 @@ shcore::Value CollectionModify::array_append(const shcore::Argument_list &args)
 *
 * The attribute deletion will be done on the collection's documents once the execute method is called.
 *
-* This function can be invoked after:
+* This function can be invoked multiple times after:
 *
 * - modify(String searchCondition)
 * - set(String attribute, Value value)
@@ -516,13 +509,12 @@ shcore::Value CollectionModify::array_append(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::arrayDelete(String path, Value value)
-{}
+CollectionModify CollectionModify::arrayDelete(String path, Value value){}
 #endif
 shcore::Value CollectionModify::array_delete(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(1, "CollectionModify::arrayDelete");
+  args.ensure_count(1, "CollectionModify.arrayDelete");
 
   try
   {
@@ -531,7 +523,7 @@ shcore::Value CollectionModify::array_delete(const shcore::Argument_list &args)
     // Updates the exposed functions
     update_functions("operation");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::arrayDelete");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.arrayDelete");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -547,7 +539,7 @@ shcore::Value CollectionModify::array_delete(const shcore::Argument_list &args)
 *
 * This method is usually used in combination with limit to fix the amount of documents to be updated.
 *
-* This function can be invoked after:
+* This function can be invoked only once after:
 *
 * - set(String attribute, Value value)
 * - unset(String attribute)
@@ -564,12 +556,11 @@ shcore::Value CollectionModify::array_delete(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::sort(List sortExprStr)
-{}
+CollectionModify CollectionModify::sort(List sortExprStr){}
 #endif
 shcore::Value CollectionModify::sort(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "CollectionModify::sort");
+  args.ensure_count(1, "CollectionModify.sort");
 
   try
   {
@@ -584,7 +575,7 @@ shcore::Value CollectionModify::sort(const shcore::Argument_list &args)
 
     update_functions("sort");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::sort");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.sort");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -597,7 +588,7 @@ shcore::Value CollectionModify::sort(const shcore::Argument_list &args)
 *
 * This method is usually used in combination with sort to fix the amount of documents to be updated.
 *
-* This function can be invoked after:
+* This function can be invoked only once after:
 *
 * - set(String attribute, Value value)
 * - unset(String attribute)
@@ -614,12 +605,11 @@ shcore::Value CollectionModify::sort(const shcore::Argument_list &args)
 *
 * \sa Usage examples at execute(ExecuteOptions options).
 */
-CollectionModify CollectionModify::limit(Integer numberOfDocs)
-{}
+CollectionModify CollectionModify::limit(Integer numberOfDocs){}
 #endif
 shcore::Value CollectionModify::limit(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "CollectionModify::limit");
+  args.ensure_count(1, "CollectionModify.limit");
 
   try
   {
@@ -627,14 +617,14 @@ shcore::Value CollectionModify::limit(const shcore::Argument_list &args)
 
     update_functions("limit");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify::limit");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("CollectionModify.limit");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 shcore::Value CollectionModify::bind(const shcore::Argument_list &UNUSED(args))
 {
-  throw shcore::Exception::logic_error("CollectionModify::bind: not yet implemented.");
+  throw shcore::Exception::logic_error("CollectionModify.bind: not yet implemented.");
 
   return Value(Object_bridge_ref(this));
 }
@@ -688,12 +678,11 @@ shcore::Value CollectionModify::bind(const shcore::Argument_list &UNUSED(args))
 * var res_car = collection.modify().arrayDelete('hobbies[0]').sort(['age']).limit(1).execute();*
 * \endcode
 */
-Collection_resultset CollectionModify::execute(ExecuteOptions opt)
-{}
+Collection_resultset CollectionModify::execute(ExecuteOptions opt){}
 #endif
 shcore::Value CollectionModify::execute(const shcore::Argument_list &args)
 {
-  args.ensure_count(0, "CollectionModify::execute");
+  args.ensure_count(0, "CollectionModify.execute");
 
   return shcore::Value::wrap(new mysqlx::Collection_resultset(boost::shared_ptr< ::mysqlx::Result>(_modify_statement->execute())));
 }
