@@ -40,11 +40,10 @@ namespace mysh
     class Schema;
 
     /**
-    * Represents a Table on an Schema, retrieved with a session created using the X Protocol.
+    * Represents a Table on an Schema, retrieved with a session created using mysqlx module.
     * \todo Implement and document as()
     * \todo Implement and document getIndexes()
     * \todo Implement and document count()
-    * \todo Implement and document drop()
     */
     class Table : public DatabaseObject, public boost::enable_shared_from_this<Table>
     {
@@ -56,10 +55,10 @@ namespace mysh
       virtual std::string class_name() const { return "Table"; }
 #ifdef DOXYGEN
       TableInsert insert();
-      TableInsert insert([field, field, ...]);
-      TableInsert insert({ field:value, field : value, ... });
+      TableInsert insert(List columns);
+      TableInsert insert(String col1, String col2, ...);
       TableSelect select();
-      TableSelect select([field, field, ...]);
+      TableSelect select(List columns);
       TableUpdate update();
       TableDelete delete();
 #endif
