@@ -56,44 +56,30 @@ TableSelect::TableSelect(boost::shared_ptr<Table> owner)
 
 #ifdef DOXYGEN
 /**
-* Creates a new select statement and returns it.
-* The method must be executed first. After it is invoked, the following methods can be invoked: where, groupBy, orderBy, limit, bind, execute.
-* \sa where(), groupBy(), orderBy(), limit(), bind(), execute().
-* \param [field, field, ...], a optional array of Strings representing field names, which will become the projection (the columns in the result set) of the select statement. 
-*   If no field list is provided then the select returns all the fields in the underlying table.
-* \return a new TableSelect object.
-* \code{.js}
-* // open a connection
-* var mysqlx = require('mysqlx').mysqlx;
-* var mysession = mysqlx.getNodeSession("root:123@localhost:33060");
-* // create some initial data
-* mysession.sql('drop schema if exists js_shell_test;').execute();
-* mysession.sql('create schema js_shell_test;').execute();
-* mysession.sql('use js_shell_test;').execute();
-* mysession.sql('create table table1 (name varchar(50), age integer, gender varchar(20));').execute();
-* // create some initial data, populate table
-* var schema = mysession.getSchema('js_shell_test');
-* var table = schema.getTable('table1');
-* var result = table.insert({name: 'jack', age: 17, gender: 'male'}).execute();
-* var result = table.insert({name: 'adam', age: 15, gender: 'male'}).execute();
-* var result = table.insert({name: 'brian', age: 14, gender: 'male'}).execute();
-* // check results
-* table.select().execute();
-* // delete a record
-* var crud = table.delete();
-* crud.where('age = 15');
-* crud.execute();
-* // check results again
-* table.select().execute();
-* \endcode
+* Initializes this record selection handler.
+* \param searchExprStr: An optional list of string expressions identifying the columns to be retrieved, alias support is enabled on these fields.
+* \return This TableSelect object.
+*
+* The TableSelect handler will only retrieve the columns that were included on the filter, if no filter was set then all the columns will be included.
+*
+* Calling this function is allowed only for the first time and it is done automatically when Table.select() is called.
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - where(String searchCriteria)
+* - groupBy(List searchExprStr)
+* - orderBy(List sortExprStr);
+* - limit(Integer numberOfRows);
+* - execute(ExecuteOptions options), including usage examples.
+*
+* \sa Usage examples at execute(ExecuteOptions options).
 */
-TableSelect TableSelect::select([field, field, ...])
-{}
+TableSelect TableSelect::select(List searchExprStr){}
 #endif
 shcore::Value TableSelect::select(const shcore::Argument_list &args)
 {
   // Each method validates the received parameters
-  args.ensure_count(0, 1, "TableSelect::select");
+  args.ensure_count(0, 1, "TableSelect.select");
 
   boost::shared_ptr<Table> table(boost::static_pointer_cast<Table>(_owner.lock()));
 
@@ -116,7 +102,7 @@ shcore::Value TableSelect::select(const shcore::Argument_list &args)
       // Updates the exposed functions
       update_functions("select");
     }
-    CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::select");
+    CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.select");
   }
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
@@ -124,19 +110,28 @@ shcore::Value TableSelect::select(const shcore::Argument_list &args)
 
 #ifdef DOXYGEN
 /**
-* Sets the expression with the search criteria, those rows that met it will be retrieved in the result.
-* This method can be invoked after the select method.
-* After this method invocation the following methods can be invoked: groupBy, orderBy, limit, bind, execute.
-* \sa select(), groupBy(), orderBy(), limit(), bind(), execute().
-* \param searchCriteria an String with the expression criteria to use to filter out those rows that don't meet it.
-* \return the same TableSelect object where the method was applied.
-*/
-TableSelect TableSelect::where(searchCriteria)
-{}
+* Sets the search condition to filter the records to be retrieved from the owner Table.
+* \param searchCondition: An optional expression to filter the records to be retrieved;
+* if not specified all the records will be retrieved from the table unless a limit is set.
+* \return This TableSelect object.
+*
+* This function can be invoked only once after:
+*
+* - select(List columns)
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - groupBy(List searchExprStr)
+* - orderBy(List sortExprStr)
+* - limit(Integer numberOfRows)
+* - execute(ExecuteOptions options).
+*
+* \sa Usage examples at execute(ExecuteOptions options).
+*/TableSelect TableSelect::where(String searchCondition){}
 #endif
 shcore::Value TableSelect::where(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::where");
+  args.ensure_count(1, "TableSelect.where");
 
   try
   {
@@ -144,27 +139,36 @@ shcore::Value TableSelect::where(const shcore::Argument_list &args)
 
     update_functions("where");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::where");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.where");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
 /**
-* Sets a grouping criteria for the resultset, if used, the TableSelect operation will group the records using the stablished criteria.
-* This method can be invoked after the following method have been invoked: where.
-* This method can be invoked after the methods find and fields. And after method groupBy invocation the following methods can be invoked: having, sort, limit, bind, execute.
-*
-* \sa select(), where(), having(), sort(), limit(), bind(), execute().
+* Sets a grouping criteria for the resultset.
 * \param searchExprStr: A list of string expressions identifying the grouping criteria.
-* \return the same TableSelect object where the method was applied.
-*/
-TableSelect TableSelect::groupBy(List searchExprStr)
-{}
+* \return This TableSelect object.
+*
+* If used, the TableSelect handler will group the records using the stablished criteria.
+*
+* This function can be invoked only once after:
+* - select(List projectedSearchExprStr)
+* - where(String searchCondition)
+*
+* After this function invocation the following functions can be invoked:
+*
+* - having(String searchCondition)
+* - orderBy(List sortExprStr);
+* - limit(Integer numberOfRows);
+* - execute(ExecuteOptions options).
+*
+* \sa Usage examples at execute(ExecuteOptions options).
+*/TableSelect TableSelect::groupBy(List searchExprStr){}
 #endif
 shcore::Value TableSelect::group_by(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::groupBy");
+  args.ensure_count(1, "TableSelect.groupBy");
 
   try
   {
@@ -179,27 +183,37 @@ shcore::Value TableSelect::group_by(const shcore::Argument_list &args)
 
     update_functions("groupBy");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::groupBy");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.groupBy");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
 /**
-* Sets a condition for records to be considered in agregate function operations, if used the CollectionFind operation will only consider the records matching the stablished criteria.
-* Calling this method is allowed only for the first time and only if the grouping criteria has been set by calling TableSelect.groupBy(groupCriteria), after that its usage is forbidden since the internal class state has been updated to handle the rest of the Select operation.
-* After this method invocation the following methods can be invoked: sort, limit, bind, execute.
-*
-* \sa groupBy(), sort(), limit(), bind(), execute().
+* Sets a condition for records to be considered in agregate function operations.
 * \param searchCondition: A condition on the agregate functions used on the grouping criteria.
-* \return the same TableSelect object where the method was applied.
+* \return This TableSelect object.
+*
+* If used the TableSelect operation will only consider the records matching the stablished criteria.
+*
+* This function can be invoked only once after:
+*
+* - groupBy(List searchExprStr)
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - orderBy(List sortExprStr)
+* - limit(Integer numberOfRows)
+* - execute(ExecuteOptions options).
+*
+* \sa Usage examples at execute(ExecuteOptions options).
 */
-TableSelect TableSelect::having(String searchCondition)
-{}
+
+TableSelect TableSelect::having(String searchCondition){}
 #endif
 shcore::Value TableSelect::having(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::having");
+  args.ensure_count(1, "TableSelect.having");
 
   try
   {
@@ -207,27 +221,41 @@ shcore::Value TableSelect::having(const shcore::Argument_list &args)
 
     update_functions("having");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::having");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.having");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
 /**
-* Sets the sorting criteria to be used on the Resultset, if used the TableSelect operation will return the records sorted with the defined criteria.
-* Calling this method is allowed only for the first time and only if the search criteria has been set by calling TableSelect.where(), after that its usage is forbidden since the internal class state has been updated to handle the rest of the TableSelect operation.
-* And after this method invocation the following methods can be invoked: limit, bind, execute.
+* Sets the sorting criteria to be used on the Resultset.
+* \param sortExprStr: A list of expression strings defining the sort criteria for the returned records.
+* \return This TableSelect object.
 *
-* \sa limit(), bind(), execute()
-* \param sortExprStr: A list containing the sort criteria expressions to be used on the operation.
-* \return the same TableSelect object where the method was applied.
+* If used the TableSelect handler will return the records sorted with the defined criteria.
+*
+* The elements of sortExprStr list are strings defining the column name on which the sorting will be based in the form of "columnIdentifier [ ASC | DESC ]".
+* If no order criteria is specified, ascending will be used by default.
+*
+* This function can be invoked only once after:
+*
+* - select(List projectedSearchExprStr)
+* - where(String searchCondition)
+* - groupBy(List searchExprStr)
+* - having(String searchCondition)
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - limit(Integer numberOfRows)
+* - execute(ExecuteOptions options).
+*
+* \sa Usage examples at execute(ExecuteOptions options).
 */
-TableSelect TableSelect::orderBy(List sortExprStr)
-{}
+TableSelect TableSelect::orderBy(List sortExprStr){}
 #endif
 shcore::Value TableSelect::order_by(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::orderBy");
+  args.ensure_count(1, "TableSelect.orderBy");
 
   try
   {
@@ -242,27 +270,39 @@ shcore::Value TableSelect::order_by(const shcore::Argument_list &args)
 
     update_functions("orderBy");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::orderBy");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.orderBy");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
 /**
-* Sets the maximum number of documents to be returned on the select operation, if used the TableSelect operation will return at most numberOfRows documents.
-* Calling this method is allowed only for the first time and only if the search criteria has been set by calling TableSelect.where(searchCriteria), after that its usage is forbidden since the internal class state has been updated to handle the rest of the TableSelect operation.
-* After this method invocation the following methods can be invoked: offset, bind, execute.
+* Sets the maximum number of records to be returned on the select operation.
+* \param numberOfRows: The maximum number of records to be retrieved.
+* \return This TableSelect object.
 *
-* \sa offset(), bind(), execute().
-* \param numberOfRows: The maximum number of documents to be retrieved.
-* \return the same TableSelect object where the method was applied.
+* If used, the TableSelect operation will return at most numberOfRows records.
+*
+* This function can be invoked only once after:
+*
+* - select(List projectedSearchExprStr)
+* - where(String searchCondition)
+* - groupBy(List searchExprStr)
+* - having(String searchCondition)
+* - orderBy(List sortExprStr)
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - offset(Integer limitOffset)
+* - execute(ExecuteOptions options).
+*
+* \sa Usage examples at execute(ExecuteOptions options).
 */
-TableSelect TableSelect::limit(Integer numberOfRows)
-{}
+TableSelect TableSelect::limit(Integer numberOfRows){}
 #endif
 shcore::Value TableSelect::limit(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::limit");
+  args.ensure_count(1, "TableSelect.limit");
 
   try
   {
@@ -270,7 +310,7 @@ shcore::Value TableSelect::limit(const shcore::Argument_list &args)
 
     update_functions("limit");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::limit");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.limit");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
@@ -278,19 +318,24 @@ shcore::Value TableSelect::limit(const shcore::Argument_list &args)
 #ifdef DOXYGEN
 /**
 * Sets number of records to skip on the resultset when a limit has been defined.
-* Calling this method is allowed only for the first time and only if a limit has been set by calling TableSelect.limit(numberOfRows), after that its usage is forbidden since the internal class state has been updated to handle the rest of the TableSelect operation.
-* After this method, the following methods can be invoked: bind, execute.
+* \param limitOffset: The number of records to skip before start including them on the Resultset.
+* \return This TableSelect object.
 *
-* \sa limit(), bind(), execute()
-* \param limitOffset: The number of documents to skip before start including them on the Resultset.
-* \return the same TableSelect object where the method was applied.
+* This function can be invoked only once after:
+*
+* - limit(Integer numberOfRows)
+*
+* After this function invocation, the following functions can be invoked:
+*
+* - execute(ExecuteOptions options)
+*
+* \sa Usage examples at execute(ExecuteOptions options).
 */
-TableSelect TableSelect::offset(Integer limitOffset)
-{}
+TableSelect TableSelect::offset(Integer limitOffset){}
 #endif
 shcore::Value TableSelect::offset(const shcore::Argument_list &args)
 {
-  args.ensure_count(1, "TableSelect::offset");
+  args.ensure_count(1, "TableSelect.offset");
 
   try
   {
@@ -298,39 +343,84 @@ shcore::Value TableSelect::offset(const shcore::Argument_list &args)
 
     update_functions("offset");
   }
-  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect::offset");
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.offset");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
-/**
-* Sets the bindings mappings for the TableSelect.
-* \param { var:val, var : val, ... } a Map a map of key value pairs where each key is the name of a parameter and each value is the value of the parameter.
-* \return the same TableSelect object where the method was applied.
-*/
 TableSelect TableSelect::bind({ var:val, var : val, ... })
 {}
 #endif
 shcore::Value TableSelect::bind(const shcore::Argument_list &UNUSED(args))
 {
-  throw shcore::Exception::logic_error("TableSelect::bind: not yet implemented.");
+  throw shcore::Exception::logic_error("TableSelect.bind: not yet implemented.");
 
   return Value(boost::static_pointer_cast<Object_bridge>(shared_from_this()));
 }
 
 #ifdef DOXYGEN
 /**
-* Executes the TableSelect operation with all the configured options and returns.
-* \param opt the execution options (currently not used).
-* \return a ResultSet object that can be used to retrieve the results.
-*/
-ResultSet TableSelect::execute(ExecuteOptions opt)
-{}
+* Executes the Find operation with all the configured options and returns.
+* \return Collection_resultset A Collection resultset object that can be used to retrieve the results of the find operation.
+*
+* This function can be invoked after any other function on this class.
+*
+* Examples:
+* \code{.js}
+* // open a connection
+* var mysqlx = require('mysqlx').mysqlx;
+* var mysession = mysqlx.getSession("myuser@localhost", mypwd);
+*
+* // Creates a table named friends on the test schema
+* mysession.sql('create table test.friends (name varchar(50), age integer, gender varchar(20));').execute();
+*
+* var table = mysession.test.friends;
+*
+* // create some initial data
+* table.insert('name','last_name','age','gender')
+*      .values('jack','black', 17, 'male')
+* .    .values('adam', 'sandler', 15, 'male')
+* .    .values('brian', 'adams', 14, 'male')
+* .    .values('alma', 'lopez', 13, 'female').execute();
+
+* table.insert(['name','last_name','age','gender'])
+* .    .values('carol', 'shiffield', 14, 'female')
+* .    .values('donna', 'summers', 16, 'female')
+* .    .values('angel', 'down', 14, 'male').execute();
+
+*
+* // Retrieve all the records from the table
+* var res_all = table.find().execute();
+*
+* // Retrieve the records for all males
+* var res_males = table.select().where('gender="male"').execute();
+*
+* // Retrieve the name and last name only
+* var res_partial = table.select(['name', 'last_name']).execute();
+*
+* // Retrieve the records sorted by age in descending order
+* var res_sorted = table.select(['name', 'last_name', 'age']).orderBy(['age desc']).execute();
+*
+* // Retrieve the four younger friends
+* var res_youngest = table.select().orderBy(['age']).limit(4).execute();
+*
+* // Retrieve the four younger friends after the youngest
+* var res = table.select().orderBy(['age']).limit(4).offset(1).execute();
+* \endcode
+*/ResultSet TableSelect::execute(ExecuteOptions options){}
 #endif
 shcore::Value TableSelect::execute(const shcore::Argument_list &args)
 {
-  args.ensure_count(0, "TableSelect::execute");
+  mysqlx::Resultset *result = NULL;
 
-  return shcore::Value::wrap(new mysqlx::Resultset(boost::shared_ptr< ::mysqlx::Result>(_select_statement->execute())));
+  try
+  {
+    args.ensure_count(0, "TableSelect.execute");
+
+    result = new mysqlx::Resultset(boost::shared_ptr< ::mysqlx::Result>(_select_statement->execute()));
+  }
+  CATCH_AND_TRANSLATE_CRUD_EXCEPTION("TableSelect.execute");
+
+  return result ? shcore::Value::wrap(result) : shcore::Value::Null();
 }
