@@ -96,7 +96,7 @@ namespace shcore {
 
     exec_and_out_equals("var table_schema = table.getSchema();");
 
-    exec_and_out_equals("print(table_schema)", "<Schema:js_shell_test>");
+    exec_and_out_equals("print(table_schema)", "<ClassicSchema:js_shell_test>");
   }
 
   // Tests table.schema
@@ -106,6 +106,26 @@ namespace shcore {
 
     exec_and_out_equals("var table = session.js_shell_test.table1;");
 
-    exec_and_out_equals("print(table.schema)", "<Schema:js_shell_test>");
+    exec_and_out_equals("print(table.schema)", "<ClassicSchema:js_shell_test>");
+  }
+
+  // Tests table.drop() and table.existInDatabase()
+  TEST_F(Shell_js_mysql_table_tests, mysql_table_drop_exist_in_database)
+  {
+    exec_and_out_equals("var schema = session.createSchema('my_sample_schema');");
+
+    exec_and_out_equals("session.sql('create table my_sample_schema.my_sample_table (name varchar(50));');");
+
+    exec_and_out_equals("var table = schema.my_sample_table;");
+
+    exec_and_out_equals("print(table.existInDatabase());", "true");
+
+    exec_and_out_equals("table.drop();");
+
+    exec_and_out_equals("print(table.existInDatabase());", "false");
+
+    exec_and_out_equals("schema.drop();");
+
+    exec_and_out_equals("session.close();");
   }
 }
