@@ -60,22 +60,7 @@ namespace shcore {
     std::string uri = mysh::strip_password(_uri);
 
     exec_and_out_equals("var session = mysqlx.getSession('" + _uri + "');");
-    exec_and_out_equals("print(session);", "<Session:" + uri + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 10)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[8] == 'schemas');", "true");
-    exec_and_out_equals("print(members[9] == 'uri');", "true");
-
+    exec_and_out_equals("print(session);", "<XSession:" + uri + ">");
     exec_and_out_equals("session.close();");
   }
 
@@ -94,22 +79,7 @@ namespace shcore {
       password = _pwd;
 
     exec_and_out_equals("var session = mysqlx.getSession('" + _uri + "', '" + password + "');");
-    exec_and_out_equals("print(session);", "<Session:" + uri + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 10)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[8] == 'schemas');", "true");
-    exec_and_out_equals("print(members[9] == 'uri');", "true");
-
+    exec_and_out_equals("print(session);", "<XSession:" + uri + ">");
     exec_and_out_equals("session.close();");
   }
 
@@ -138,22 +108,7 @@ namespace shcore {
     uri << user << "@" << host << ":" << port;
 
     exec_and_out_equals("var session = mysqlx.getSession(" + connection_data.str() + ");");
-    exec_and_out_equals("print(session);", "<Session:" + uri.str() + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 10)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[8] == 'schemas');", "true");
-    exec_and_out_equals("print(members[9] == 'uri');", "true");
-
+    exec_and_out_equals("print(session);", "<XSession:" + uri.str() + ">");
     exec_and_out_equals("session.close();");
   }
 
@@ -181,22 +136,7 @@ namespace shcore {
     uri << user << "@" << host << ":" << port;
 
     exec_and_out_equals("var session = mysqlx.getSession(" + connection_data.str() + ", '" + password + "');");
-    exec_and_out_equals("print(session);", "<Session:" + uri.str() + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 10)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[8] == 'schemas');", "true");
-    exec_and_out_equals("print(members[9] == 'uri');", "true");
-
+    exec_and_out_equals("print(session);", "<XSession:" + uri.str() + ">");
     exec_and_out_equals("session.close();");
   }
 
@@ -208,22 +148,6 @@ namespace shcore {
     std::string uri = mysh::strip_password(_uri);
     exec_and_out_equals("var session = mysqlx.getNodeSession('" + _uri + "');");
     exec_and_out_equals("print(session);", "<NodeSession:" + uri + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 11)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'sql');", "true");
-    exec_and_out_equals("print(members[8] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[9] == 'schemas');", "true");
-    exec_and_out_equals("print(members[10] == 'uri');", "true");
-
     exec_and_out_equals("session.close();");
   }
 
@@ -243,22 +167,6 @@ namespace shcore {
 
     exec_and_out_equals("var session = mysqlx.getNodeSession('" + _uri + "', '" + password + "');");
     exec_and_out_equals("print(session);", "<NodeSession:" + uri + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 11)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'sql');", "true");
-    exec_and_out_equals("print(members[8] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[9] == 'schemas');", "true");
-    exec_and_out_equals("print(members[10] == 'uri');", "true");
-
     exec_and_out_equals("session.close();");
   }
 
@@ -288,22 +196,6 @@ namespace shcore {
 
     exec_and_out_equals("var session = mysqlx.getNodeSession(" + connection_data.str() + ");");
     exec_and_out_equals("print(session);", "<NodeSession:" + uri.str() + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 11)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'sql');", "true");
-    exec_and_out_equals("print(members[8] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[9] == 'schemas');", "true");
-    exec_and_out_equals("print(members[10] == 'uri');", "true");
-
     exec_and_out_equals("session.close();");
   }
 
@@ -332,22 +224,6 @@ namespace shcore {
 
     exec_and_out_equals("var session = mysqlx.getNodeSession(" + connection_data.str() + ", '" + password + "');");
     exec_and_out_equals("print(session);", "<NodeSession:" + uri.str() + ">");
-
-    // Ensures the right members exist
-    exec_and_out_equals("var members = dir(session);");
-    exec_and_out_equals("print(members.length >= 11)", "true");
-    exec_and_out_equals("print(members[0] == 'close');", "true");
-    exec_and_out_equals("print(members[1] == 'getDefaultSchema');", "true");
-    exec_and_out_equals("print(members[2] == 'getSchema');", "true");
-    exec_and_out_equals("print(members[3] == 'getSchemas');", "true");
-    exec_and_out_equals("print(members[4] == 'getUri');", "true");
-    exec_and_out_equals("print(members[5] == 'setDefaultSchema');", "true");
-    exec_and_out_equals("print(members[6] == 'setFetchWarnings');", "true");
-    exec_and_out_equals("print(members[7] == 'sql');", "true");
-    exec_and_out_equals("print(members[8] == 'defaultSchema');", "true");
-    exec_and_out_equals("print(members[9] == 'schemas');", "true");
-    exec_and_out_equals("print(members[10] == 'uri');", "true");
-
     exec_and_out_equals("session.close();");
   }
 }
