@@ -123,34 +123,34 @@ namespace shcore {
     {
       SCOPED_TRACE("Testing parameter validation on delete");
       exec_and_out_equals("table.delete();");
-      exec_and_out_contains("table.delete(5);", "", "Invalid number of arguments in TableDelete::delete, expected 0 but got 1");
+      exec_and_out_contains("table.delete(5);", "", "Invalid number of arguments in TableDelete.delete, expected 0 but got 1");
     }
 
     {
       SCOPED_TRACE("Testing parameter validation on where");
-      exec_and_out_contains("table.delete().where();", "", "Invalid number of arguments in TableDelete::where, expected 1 but got 0");
-      exec_and_out_contains("table.delete().where(5);", "", "TableDelete::where: Argument #1 is expected to be a string");
-      exec_and_out_contains("table.delete().where('test = \"2');", "", "TableDelete::where: Unterminated quoted string starting at 8");
+      exec_and_out_contains("table.delete().where();", "", "Invalid number of arguments in TableDelete.where, expected 1 but got 0");
+      exec_and_out_contains("table.delete().where(5);", "", "TableDelete.where: Argument #1 is expected to be a string");
+      exec_and_out_contains("table.delete().where('test = \"2');", "", "TableDelete.where: Unterminated quoted string starting at 8");
       exec_and_out_equals("table.delete().where('test = \"2\"');");
     }
 
     {
       SCOPED_TRACE("Testing parameter validation on orderBy");
-      exec_and_out_contains("table.delete().orderBy();", "", "Invalid number of arguments in TableDelete::orderBy, expected 1 but got 0");
-      exec_and_out_contains("table.delete().orderBy(5);", "", "TableDelete::orderBy: Argument #1 is expected to be an array");
-      exec_and_out_contains("table.delete().orderBy([]);", "", "TableDelete::orderBy: Order criteria can not be empty");
-      exec_and_out_contains("table.delete().orderBy(['test', 5]);", "", "TableDelete::orderBy: Element #2 is expected to be a string");
+      exec_and_out_contains("table.delete().orderBy();", "", "Invalid number of arguments in TableDelete.orderBy, expected 1 but got 0");
+      exec_and_out_contains("table.delete().orderBy(5);", "", "TableDelete.orderBy: Argument #1 is expected to be an array");
+      exec_and_out_contains("table.delete().orderBy([]);", "", "TableDelete.orderBy: Order criteria can not be empty");
+      exec_and_out_contains("table.delete().orderBy(['test', 5]);", "", "TableDelete.orderBy: Element #2 is expected to be a string");
       exec_and_out_contains("table.delete().orderBy(['test']);", "", "");
     }
 
     {
       SCOPED_TRACE("Testing parameter validation on limit");
-      exec_and_out_contains("table.delete().limit();", "", "Invalid number of arguments in TableDelete::limit, expected 1 but got 0");
-      exec_and_out_contains("table.delete().limit('');", "", "TableDelete::limit: Argument #1 is expected to be an unsigned int");
+      exec_and_out_contains("table.delete().limit();", "", "Invalid number of arguments in TableDelete.limit, expected 1 but got 0");
+      exec_and_out_contains("table.delete().limit('');", "", "TableDelete.limit: Argument #1 is expected to be an unsigned int");
       exec_and_out_equals("table.delete().limit(5);");
     }
 
-    exec_and_out_contains("table.delete().bind();", "", "TableDelete::bind: not yet implemented.");
+    exec_and_out_contains("table.delete().bind();", "", "TableDelete.bind: not yet implemented.");
 
     exec_and_out_equals("session.close();");
   }
