@@ -13,6 +13,10 @@
    along with this program; if not, write to the Free Software
    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA */
 
+#ifdef HAVE_PYTHON
+#include <Python.h>
+#endif
+
 #include <gtest/gtest.h>
 #include <fstream>
 #include <iostream>
@@ -24,12 +28,19 @@ int main(int argc, char **argv)
 
   JScript_context_init();
 #endif
+
+#ifdef HAVE_PYTHON
+#ifdef _WINDOWS
+  Py_NoSiteFlag = 1;
+#endif
+  Py_InitializeEx(0);
+#endif
+
   if (!getenv("MYSQL_URI"))
   {
     std::cerr << "WARNING: The MYSQL_URI MYSQL_PWD and MYSQL_PORT environment variables are not set\n";
   }
 
-  //::testing::GTEST_FLAG(filter) = "Shell_js_crud_table_insert_tests*";
   ::testing::InitGoogleTest(&argc, argv);
 
   const char *generate_option = "--generate_test_groups=";
@@ -50,5 +61,14 @@ int main(int argc, char **argv)
     return 0;
   }
 
-  return RUN_ALL_TESTS();
+  int ret_val = RUN_ALL_TESTS();
+
+#ifdef HAVE_PYTHON
+#ifdef _WINDOWS
+  Py_NoSiteFlag = 1;
+#endif
+  Py_InitializeEx(0);
+#endif
+
+  return ret_val;
 }
